backend/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# LLM API
groq==0.13.0

# News API
newsapi-python==0.2.7
requests==2.32.5

# Environment Variables
python-dotenv==1.2.1
